pyproject.toml
[project]
name = "snowsight-cli"
version = "0.1.0"
description = "A TUI SQL client for Snowflake"
requires-python = ">=3.11"
dependencies = [
    "textual[syntax]>=0.70.0",
    "snowflake-connector-python>=3.10.0",
    "toml>=0.10.2",
]

[project.scripts]
snowsightcli = "snowsight.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["snowsight"]

[tool.uv]
dev-dependencies = []
